requirements.txt
kafka-python
paho-mqtt
pyspark==3.2.0
pandas==1.4.4
numpy==1.23.3
python-dotenv==0.21.0